handle sqlite.tar.xz and _get_metadata_mysql pseudo-private func

+++ hcaquery/__init__.py
@@ -1,5 +1,3 @@
-#import anndata as ad
-#from scipy.sparse import csr_matrix
 import tempfile
 import os
 import anndata as ad
@@ -13,15 +11,16 @@ try:
 except:
 	from sqlalchemy.engine import URL
 
-#REMOTE_URL = "https://swift.rc.nectar.org.au/v1/AUTH_06d6e008e3e642da99d806ba3ea629c5/harmonised-human-atlas"
-REMOTE_URL = 'file:///vast/projects/human_cell_atlas_py' # for testing
-METADATADB_URL = URL.create("mysql", \
+REMOTE_URL = "https://swift.rc.nectar.org.au/v1/AUTH_06d6e008e3e642da99d806ba3ea629c5"
+ASSAY_URL= '{}/harmonised-human-atlas'.format(REMOTE_URL)
+METADATASQLITE_URL = '{}/metadata-sqlite/metadata.tar.xz'.format(REMOTE_URL)
+METADATADB_URL = URL.create("mysql+pymysql", \
 	username='public_access', \
 	password='password', \
-	host='7b4abe7csjh.db.cloud.edu.au', \
+	host='7f7wu64bnzu.db.cloud.edu.au', \
 	database='metadata')
 
-assay_map = {'counts': 'splitted_DB2_anndata', 'cpm': 'splitted_DB2_anndata_scaled'} #{'counts': 'original', 'cpm': 'cpm'}
+assay_map = {'counts': 'original', 'cpm': 'cpm'}
 
 #using tmpdir as cachedir for the time being
 def get_default_cache_dir():
@@ -38,30 +37,31 @@ def sync_remote_file(full_url, output_file):
 			os.makedirs(output_dir)
 		print("Downloading {url} to {outfile}".format(url=full_url, outfile=output_file))
 
-		#TODO: produce warning if file isn't found (but don't stop)
 		try:
 			import urllib.request
 			r = urllib.request.urlopen(full_url)
 			urllib.request.urlretrieve(full_url, output_file, show_progress)
 
 		except Exception as e:
-			print("File {url} could not be downloaded.".format(url=full_url))
-			print(e)
-			os.remove(output_file)
+			if os.path.isfile(output_file): 
+				os.remove(output_file)
+			import sys
+			sys.exit("File {url} could not be downloaded.\n{e}".format(url=full_url, e=e))
 
 # function to get metadata
-def get_metadata(connection_url=METADATADB_URL, \
-	localsqlite = False, cache_dir = get_default_cache_dir()):
+def get_metadata(sqlite_url=METADATASQLITE_URL, \
+	cache_dir = get_default_cache_dir()):
 
-	# branch based on whether local sqlite or remote mysql/postgres is used.
-	if localsqlite:
-		sqlite_path = os.path.join(cache_dir, "metadata.sqlite")
-		sync_remote_file(
-			'file:///vast/projects/RCP/human_cell_atlas',
-			sqlite_path)
-		eng = sqlalchemy.create_engine('sqlite:///{}'.format(sqlite_path))
-	else:
-		eng = sqlalchemy.create_engine(connection_url)
+	import tarfile
+
+	sqlite_tarxz_path = os.path.join(cache_dir, "metadata.tar.xz")
+	sqlite_path = os.path.join(cache_dir, "metadata.sqlite")
+	sync_remote_file(sqlite_url, sqlite_tarxz_path)
+	if (not os.path.isfile(sqlite_path)) and os.path.isfile(sqlite_tarxz_path):
+		with tarfile.open(sqlite_tarxz_path) as f:
+			f.extractall(cache_dir)
+
+	eng = sqlalchemy.create_engine('sqlite:///{}'.format(sqlite_path))
 
 	# get metadata table to return to user
 	md = sqlalchemy.MetaData()
@@ -70,7 +70,17 @@ def get_metadata(connection_url=METADATADB_URL, \
 
 	return eng, mdtab
 
-def sync_assay_files(url = REMOTE_URL, cache_dir = get_default_cache_dir(), subdirs = [], files = []):
+def _get_metadata_mysql(connection_url=METADATADB_URL, \
+	cache_dir = get_default_cache_dir()):
+
+	eng = sqlalchemy.create_engine(connection_url)
+	md = sqlalchemy.MetaData()
+	mdtab = sqlalchemy.Table("metadata", md, autoload_with=eng)
+	mdtab = md.tables['metadata']
+
+	return eng, mdtab
+
+def sync_assay_files(url = ASSAY_URL, cache_dir = get_default_cache_dir(), subdirs = [], files = []):
 
 	urls = ["{baseurl}/{subdir}/{samplefile}".format(baseurl=url, subdir=sdir, samplefile=ifile) 
 		for sdir, ifile in itertools.product(subdirs, files)]
@@ -105,7 +115,7 @@ def get_SingleCellExperiment(
 	data = pd.DataFrame(),
 	assays = ["counts", "cpm"],
 	cache_directory = get_default_cache_dir(),
-	repository = REMOTE_URL,
+	repository = ASSAY_URL,
 	features = [],
 	silent = False
 ):
@@ -128,7 +138,7 @@ def get_SingleCellExperiment(
 	# mapping requested assay to subdirectory name
 	subdirs = [assay_map[a] for a in assays]
 
-	sync_assay_files(cache_dir='/vast/scratch/users/yang.e/tmp/dummy-hca',subdirs=subdirs, files=files_to_read)
+	sync_assay_files(url=repository, cache_dir=cache_directory ,subdirs=subdirs, files=files_to_read)
 
 	# "outer" product of subdirectories and filenames
 	files2pull = [os.path.join(cache_directory,s,f) for s,f in itertools.product(subdirs, files_to_read)]
@@ -148,24 +158,22 @@ def get_SingleCellExperiment(
 
 if __name__=="__main__":
 
-	eng, metadatatab = get_metadata(cache_dir='/vast/projects/RCP/human_cell_atlas', localsqlite=True)
-	#sync_assay_files(cache_dir='/vast/scratch/users/yang.e/tmp/dummy-hca', subdirs=['original','cpm'], files=['18b89f46a7bd507ac85033d3e21c56d6','18ce8f08b718573c04d26094071e1e69'])
-	#df = pd.read_sql("SELECT * FROM metadata WHERE ethnicity='African';", eng)
-	#q = sqlalchemy.select(metadatatab).where(metadatatab.c.ethnicity=='African')
-	#q = sqlalchemy.select().where(\
-	#	(metadatatab.c.ethnicity=="African") & \
-	#	(metadatatab.c.assay.like('%{}%'.format('10x'))) & \
-	#	(metadatatab.c.tissue=="lung parenchyma") & \
-	#	(metadatatab.c.cell_type.like('%{}%'.format('CD4'))))
-	q = sqlalchemy.select([metadatatab.c.file_id_db, metadatatab.c.assay, metadatatab.c.tissue, metadatatab.c.cell_type]).where(\
-			(metadatatab.c.assay.like('%{}%'.format('10x'))) & \
-			(metadatatab.c.tissue=="lung parenchyma") & \
-			(metadatatab.c.cell_type.like('%{}%'.format('CD4'))))
+	eng, metadatatab = get_metadata()
+
+	q = sqlalchemy.select('*').where( \
+		metadatatab.c.ethnicity=="African", \
+		metadatatab.c.assay.like('%10x%'), \
+		metadatatab.c.tissue == "lung parenchyma", \
+		metadatatab.c.cell_type.like('%CD4%'))
+	#q = sqlalchemy.select([metadatatab.c.file_id_db, metadatatab.c.assay, metadatatab.c.tissue, metadatatab.c.cell_type]).where(\
+	#		(metadatatab.c.assay.like('%{}%'.format('10x'))) & \
+	#		(metadatatab.c.tissue=="lung parenchyma") & \
+	#		(metadatatab.c.cell_type.like('%{}%'.format('CD4'))))
 
 	with eng.connect() as con:
-		df = pd.DataFrame(con.execute(q))
+		df = pd.DataFrame(con.execute(q))	
 	eng.dispose()
 	print(df)
-	#z = get_SingleCellExperiment(df, cache_directory='/vast/projects/human_cell_atlas_py',('))
-	z = get_SingleCellExperiment(df, cache_directory='/vast/projects/human_cell_atlas_py')
+
+	z = get_SingleCellExperiment(df, repository='file:///vast/projects/human_cell_atlas_py/anndata')
 	print(z)
